Add JSON Webtokens
Authenticating using a JSON Webtoken is now supported in the SDK. Weaver.signIn(username, password) still works, but you will also be able to supply a token as the argument for signIn (Weaver.signIn(token)). If the token is still valid, the user will be signed in.

See code for changes